rename src to fix openbb command
