+++ openbb_terminal/openbb_terminal/argparse_translator/argparse_class_processor.py
@@ -4,7 +4,7 @@ from typing import Any, Dict, Optional, Type
 # TODO: this needs to be done differently
 from openbb_core.app.static.container import Container
 
-from src.argparse_translator.argparse_translator import (
+from openbb_terminal.argparse_translator.argparse_translator import (
     ArgparseTranslator,
     ReferenceToCustomArgumentsProcessor,
 )
