+++ openbb_terminal/openbb_terminal/argparse_translator/obbject_registry.py
