+++ openbb_terminal/openbb_terminal/assets/routines/routine_example.openbb
