+++ openbb_terminal/openbb_terminal/assets/styles/default/Consolas.ttf
