+++ openbb_terminal/openbb_terminal/assets/styles/default/dark.mplrc.json
