+++ openbb_terminal/openbb_terminal/assets/styles/default/light.mpfstyle.json
