+++ openbb_terminal/openbb_terminal/assets/styles/default/light.mplrc.json
