+++ openbb_terminal/openbb_terminal/assets/styles/default/light.pltstyle.json
