+++ openbb_terminal/openbb_terminal/assets/styles/default/tables.pltstyle.json
