+++ openbb_terminal/openbb_terminal/assets/styles/user/openbb.richstyle.json
