+++ openbb_terminal/openbb_terminal/config/completer.py
