+++ openbb_terminal/openbb_terminal/config/menu_text.py
