+++ openbb_terminal/openbb_terminal/controllers/hub_service.py
@@ -4,12 +4,12 @@ from typing import Optional
 
 import requests
 
-from src.config.constants import (
+from openbb_terminal.config.constants import (
     CONNECTION_ERROR_MSG,
     CONNECTION_TIMEOUT_MSG,
     TIMEOUT,
 )
-from src.session import Session
+from openbb_terminal.session import Session
 
 # created dictionaries for personal and default routines with the structure
 # {"file_name" :["script","personal/default"]}
