+++ openbb_terminal/openbb_terminal/controllers/script_parser.py
@@ -6,7 +6,7 @@ from typing import Dict, List, Match, Optional, Tuple, Union
 
 from dateutil.relativedelta import relativedelta
 
-from src.session import Session
+from openbb_terminal.session import Session
 
 # pylint: disable=too-many-statements,eval-used,consider-iterating-dictionary
 # pylint: disable=too-many-branches,too-many-return-statements
