+++ openbb_terminal/openbb_terminal/session.py
@@ -10,10 +10,10 @@ from openbb_core.app.model.user_settings import UserSettings as User
 from prompt_toolkit import PromptSession
 from prompt_toolkit.history import FileHistory
 
-from src.config.console import Console
-from src.config.constants import HIST_FILE_PROMPT
-from src.config.style import Style
-from src.models.settings import Settings
+from openbb_terminal.config.console import Console
+from openbb_terminal.config.constants import HIST_FILE_PROMPT
+from openbb_terminal.config.style import Style
+from openbb_terminal.models.settings import Settings
 
 
 class Session(metaclass=SingletonMeta):
