+++ openbb_terminal/pyproject.toml
@@ -4,7 +4,7 @@ version = "1.0.0"
 description = "Investment Research for Everyone, Anywhere."
 license = "MIT"
 authors = ["OpenBB <hello@openbb.co>"]
-packages = [{ include = "src" }]
+packages = [{ include = "openbb_terminal" }]
 include = ["openbb_terminal/.env"]
 readme = "../website/pypi.md"
 homepage = "https://openbb.co"
@@ -12,7 +12,7 @@ repository = "https://github.com/OpenBB-finance/OpenBBTerminal"
 documentation = "https://docs.openbb.co/sdk"
 
 [tool.poetry.scripts]
-openbb = 'openbb_terminal.src.cli:main'
+openbb = 'openbb_terminal.cli:main'
 
 [tool.poetry.dependencies]
 python = "^3.8.1,<3.12"
